[PL-7127]: Reference an existing Azure Key Vault secret - Backend (#11185)
* Support for Referencing Azure Vault Secrets

* Dummy Commit

* Fixing Code Formatting Issue

* Renamed File

* Changed String constant of single character to char data type

* Added Unit Tests

* Added Functional Tests For Support of Referencing Azure Vault Secrets

* Added Category and Owner to Unit Test

* Update AzureVaultFunctionalTest.java

* Update AzureVaultFunctionalTest.java

* Update AzureVaultFunctionalTest.java

* Update AzureVaultFunctionalTest.java

* Run Azure Secret Manager functional Tests in Sequential Order

* Update AzureVaultFunctionalTest.java

* Update AzureVaultFunctionalTest.java

* Update AzureVaultFunctionalTest.java

* Update AzureVaultFunctionalTest.java

* Simplified the logic

* Revert "Simplified the logic"

This reverts commit 13dc99025c4e2e20bc425499079d0d961d5bee66.

* Update SecretManagementDelegateServiceImpl.java

* Fix Azure Vault Functional Test

+++ 71-rest/src/main/java/software/wings/service/impl/security/SecretManagerImpl.java
@@ -1579,12 +1579,12 @@ public class SecretManagerImpl implements SecretManager {
           }
           break;
         case AWS_SECRETS_MANAGER:
-          break;
         case CYBERARK:
+        case AZURE_VAULT:
           break;
         default:
           throw new SecretManagementException(SECRET_MANAGEMENT_ERROR,
-              "Secret path can be specified only if the secret manager is of VAULT/AWS_SECRETS_MANAGER/CYBERARK type!",
+              "Secret path can be specified only if the secret manager is of VAULT/AWS_SECRETS_MANAGER/CYBERARK/AZURE_VAULT type!",
               USER);
       }
     }
@@ -1669,10 +1669,12 @@ public class SecretManagerImpl implements SecretManager {
         }
         break;
       case AZURE_VAULT:
-        String keyName = encryptedData.getEncryptionKey();
-        AzureVaultConfig encryptionConfig =
-            azureSecretsManagerService.getEncryptionConfig(accountId, encryptedData.getKmsId());
-        azureVaultService.delete(accountId, keyName, encryptionConfig);
+        if (isEmpty(encryptedData.getPath())) {
+          String keyName = encryptedData.getEncryptionKey();
+          AzureVaultConfig encryptionConfig =
+              azureSecretsManagerService.getEncryptionConfig(accountId, encryptedData.getKmsId());
+          azureVaultService.delete(accountId, keyName, encryptionConfig);
+        }
         break;
       default:
         break;
